fix: rename folder filtersSection
